requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
fpdf==1.7.2
idna==3.11
pdf2image==1.17.0
Pillow==12.0.0
PyMuPDF==1.26.6
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.6.1
wget==3.2
